Rename DateValue.ts to date-utils
